Fix failing tests of Appointment.test.js
